ml/requirements.txt
python-dotenv
langchain
langchain-community
langchainhub
langchain-openai
beautifulsoup4
pandas
faiss-cpu
chromadb
jq
PyPDF2
nltk 
sentence-transformers
